@@ -9,7 +9,7 @@
   ],
   "type": "project",
   "require": {
-    "php": ">=8.2",
+    "php": ">=8.3",
     "ext-fileinfo": "*",
     "firebase/php-jwt": "^6.11",
     "illuminate/config": "^v12.3.0",
